+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-24
+- `14:57` -- **Added** `apps/desktopProbe/tests/quietHours/aggregate.test.ts`
 - `14:56` -- **Added** `apps/desktopProbe/src/server/quietHours/predicate.ts`
 - `14:55` -- **Added** `apps/desktopProbe/tests/quietHours/predicate.test.ts`
 - `14:53` -- **Added** `apps/desktopProbe/jest.config.ts`